Use primaryVideo and secondaryVideo props instead of mp4 and webm (#111)
@julianstein @AlexandraKlein

So i had this laying around for a while now (i actually thought i had
deleted it but turns out i saved it in a stash). I basically rename the
props `mp4` to `primaryVideoUrl` and `webm` to `secondaryVideoUrl` and
add `primaryVideoType` and `secondaryVideoType` props.

I deprecated the `mp4` and `webm` props.

The changes should be backwards compatible (no breaking changes). This
is both good and bad. Bad because the default `primaryVideoType` for
non-transparent video is still `video/mp4` when it should be
`video/webm`. We can do a major version bump later and change those
defaults.

Does this fix your problem on Tillamook, Julian?