Feature: Support for accordion behavior (#116)
### Description

This pull request enriches the `packages/disclosure` module, enabling
the Disclosure component to seamlessly operate as an Accordion with
enhanced functionality. The cornerstone of this enhancement is the
integration of the `registerDetails` function, which plays a pivotal
role in achieving the desired accordion behavior. This function
facilitates the dynamic tracking of disclosure details, paving the way
for optimal management of open and closed states.

The enhancements also encompass the introduction of a `defaultOpenIndex`
property within `DisclosureRootProps`, enabling the specification of the
default open detail index. Additionally, context management has been
meticulously optimized to ensure efficient state control and seamless
animation handling.

### Changes

- Implements the `registerDetails` function, a key driver in realizing
the accordion effect by dynamically managing disclosure details.
- Adapts the `Disclosure` component to adeptly support function
children, streamlining the passage of `registerDetails` for an improved
accordion experience.
- Enhances the `DisclosureDetailProps` interface to incorporate
properties such as `index` and `handleClick` for refined detail control.
- Optimizes state management within the `DisclosureDetails` component
through the strategic integration of context, culminating in smoother
open and closed state transitions.
- Elevates the `DisclosureSummary` component to seamlessly embrace both
props and context-based behaviors, contributing to a consistent
accordion experience.
- Streamlines context utilization and content management in the
`DisclosureContent` component for a harmonious overall implementation.


Link to example:
https://haus-core-docs-git-feature-disclosure-accodrion-madeinhaus.vercel.app/components/disclosure#usage-as-controlled-accordion